+++ students/AndrewMiotke/session09/mailroom_oo/donors_model.py
@@ -23,3 +23,19 @@ class DonorList:
             self.donor_data = {}
         else:
             self.donor_data = {d.norm_name: d for d in donors}
+
+
+    def get_donor(self, donor_name):
+        key = donor_name.strip().lower()
+
+        # return donors_list.get(key)
+        return self.donor_data.get(key)
+
+
+    def add_donor(self, donor_name):
+        donor_name = donor_name.strip()
+        donor = (donor_name, [])
+        # donors_list[donor_name.lower()] = donor
+        self.donor_data[donor_name.lower()] = donor
+
+        return donor
